fix(alert): center close button & icon triangle
